chore: reduce duplications
