+++ client/src/constants/english.ts
@@ -13,8 +13,8 @@ const enTranslations = {
   login_password_label: 'Password',
   login_password_placeholder: 'Type your password',
   login_btn_submit: 'Login',
-  login_no_account: 'Don\'t have an account yet?',
-  login_register_here: 'Create account',
+  login_account: 'Don\'t have an account yet?',
+  login_go_other: 'Create account',
   login_back_home: 'Back to home',
 
   register_title: 'Create account',
@@ -23,8 +23,8 @@ const enTranslations = {
   register_password_label: 'Password',
   register_password_placeholder: 'Type your password',
   register_btn_submit: 'Create account',
-  register_has_account: 'Already have an account?',
-  register_go_login: 'Login',
+  register_account: 'Already have an account?',
+  register_go_other: 'Login',
   register_back_home: 'Back to home',
 
   home_nav_home: 'Home',
